chore(tests): moved more helper classed to correct namespace
