+++ engine/tests/phpunit/unit/Elgg/Database/Clauses/OrderByClauseUnitTest.php
@@ -4,10 +4,8 @@ namespace Elgg\Database\Clauses;
 
 use Elgg\Database\QueryBuilder;
 use Elgg\Database\Select;
-use Elgg\Project\Paths;
+use Elgg\Helpers\Database\Clauses\CallableOrderBy;
 use Elgg\UnitTestCase;
-
-require_once Paths::elgg() . 'engine/tests/test_files/database/clauses/CallableOrderBy.php';
 
 /**
  * @group QueryBuilder
@@ -58,7 +56,7 @@ class OrderByClauseUnitTest extends UnitTestCase {
 
 		$this->qb->orderBy('alias.guid', 'asc');
 
-		$query = new OrderByClause(\CallableOrderBy::class, 'asc');
+		$query = new OrderByClause(CallableOrderBy::class, 'asc');
 
 		$qb = Select::fromTable('entities', 'alias');
 		$qb->addClause($query);
@@ -71,7 +69,7 @@ class OrderByClauseUnitTest extends UnitTestCase {
 
 		$this->qb->orderBy('alias.guid', 'asc');
 
-		$query = new OrderByClause('\CallableOrderBy::callable', 'asc');
+		$query = new OrderByClause('\Elgg\Helpers\Database\Clauses\CallableOrderBy::callable', 'asc');
 
 		$qb = Select::fromTable('entities', 'alias');
 		$qb->addClause($query);
